Updated script with check for required SSD firmware updates.
Updated script with check for required SSD firmware updates for certain MacbookAir5,1 and MacbookAir5,2 models. This will write a second boolean to ConditionalItems.plist, named 'mavericks_needs_fw_update'. A pkginfo can then use the two to determine whether Mavericks is supported and if so whether an SSD update is required.